Map source files to output files

+++ src/analyzer.rs
@@ -2,10 +2,10 @@ use futures::StreamExt;
 use std::error::Error;
 use url::Url;
 
-use crate::repo::Repo;
+use crate::{repo::Repo, Cli};
 
 #[tokio::main]
-pub(crate) async fn analyze(repos: Vec<&str>) -> Result<(), Box<dyn Error>> {
+pub(crate) async fn analyze(repos: Vec<&str>, args: &Cli) -> Result<(), Box<dyn Error>> {
     let repo_jobs = futures::stream::iter(repos.into_iter().map(|url| async move {
         let url = Url::parse(url).unwrap_or_else(|err| {
             eprintln!("Could not parse url {}: {}", url, err);
@@ -14,10 +14,17 @@ pub(crate) async fn analyze(repos: Vec<&str>) -> Result<(), Box<dyn Error>> {
 
         let repo = Repo::new(&url);
 
-        repo.clone().await;
-        repo.metrics().await;
-        repo.analyze().await;
-        repo.delete();
+        if args.analyze_only {
+            repo.analyze().await;
+        } else {
+            repo.clone().await;
+            repo.metrics().await;
+            repo.analyze().await;
+
+            if args.delete_tmp {
+                repo.delete();
+            }
+        }
     }))
     .buffer_unordered(8)
     .collect::<Vec<()>>();
